# Follow best practices for pointer aliasing in alignment inner loop.