--- a/src/cutadapt/_align.pyx
+++ b/src/cutadapt/_align.pyx
@@ -416,8 +416,8 @@ cdef class Aligner:
             # To access the diagonal cell to the upper left,
             # we store it here before overwriting it.
             _Entry diag_entry
-            _Entry current_cell
-            _Entry previous_cell
+            _Entry current_entry
+            _Entry previous_entry
 
         with nogil:
             # iterate over columns
@@ -443,11 +443,11 @@ cdef class Aligner:
                         score = diag_entry.score + match_score
                     else:
                         # Characters do not match.
-                        current_cell = column[i]
-                        previous_cell = column[i-1]
+                        current_entry = column[i]
+                        previous_entry = column[i-1]
                         cost_diag = diag_entry.cost + 1
-                        cost_deletion = current_cell.cost + deletion_cost
-                        cost_insertion = previous_cell.cost + insertion_cost
+                        cost_deletion = current_entry.cost + deletion_cost
+                        cost_insertion = previous_entry.cost + insertion_cost
 
                         if cost_diag <= cost_deletion and cost_diag <= cost_insertion:
                             # MISMATCH
@@ -457,15 +457,15 @@ cdef class Aligner:
                         elif cost_insertion <= cost_deletion:
                             # INSERTION
                             cost = cost_insertion
-                            origin = previous_cell.origin
+                            origin = previous_entry.origin
                             # penalize insertions slightly
-                            score = previous_cell.score + insertion_score
+                            score = previous_entry.score + insertion_score
                         else:
                             # DELETION
                             cost = cost_deletion
-                            origin = current_cell.origin
+                            origin = current_entry.origin
                             # penalize deletions slightly
-                            score = current_cell.score + deletion_score
+                            score = current_entry.score + deletion_score
 
                     # Remember the current cell for next iteration
                     diag_entry = column[i]