@marcelm 2023-12-12
Oh wow, very nice! Let me have a look at the compiled code before merging this.