@rhpvorderman 2023-12-12
Done! Have you managed to figure out the assembly? I cannot comprehend why it is faster. So I ran it on my laptop, a ryzen 5 5650 pro. No difference. 
Work PC with intel i7-6700: the difference above. Consistently. 
A better memory and cache loading system for the newer generation AMD cpu's, so that it cannot take as much advantage from this change?