@marcelm 2023-12-12
No, I realized I don’t have the time, so I’ll just accept the speed improvement ...